feat(android): provision and run real Google Android Emulator isolated on F drive v3.8.37

+++ AutoGram App/android/app/src/main/java/com/autogram/app/ui/settings/SettingsScreen.kt
@@ -88,7 +88,7 @@ fun SettingsScreenContent(
                             text = stringResource(
                                 R.string.settings_encoder_value,
                                 hw.bestEncoder,
-                                hw.priority
+                                hw.priority.toLong()
                             ),
                             style = MaterialTheme.typography.bodyLarge,
                             color = TextSecondaryDark
